+++ VanillaBasicsCore/src/main/kotlin/org/tobi29/scapes/vanilla/basics/material/block/BlockWater.kt
@@ -97,7 +97,7 @@ class BlockWater(type: VanillaMaterialType) : VanillaBlock(type) {
 
     override fun isSolid(data: Int) = false
 
-    override fun lightTrough(data: Int) = -1
+    override fun lightTrough(data: Int) = (-1).toByte()
 
     override fun connectStage(terrain: TerrainClient,
                               x: Int,
