+++ VanillaBasicsCore/src/main/kotlin/org/tobi29/scapes/vanilla/basics/material/block/device/BlockAlloy.kt
@@ -106,7 +106,7 @@ class BlockAlloy(type: VanillaMaterialType) : VanillaBlockContainer<EntityAlloyC
 
     override fun isTransparent(data: Int) = true
 
-    override fun lightTrough(data: Int) = -1
+    override fun lightTrough(data: Int) = (-1).toByte()
 
     override fun connectStage(terrain: TerrainClient,
                               x: Int,
