+++ VanillaBasicsCore/src/main/kotlin/org/tobi29/scapes/vanilla/basics/material/block/structural/BlockGlass.kt
@@ -61,7 +61,7 @@ class BlockGlass(type: VanillaMaterialType) : VanillaBlock(type) {
 
     override fun isTransparent(data: Int) = true
 
-    override fun lightTrough(data: Int) = -1
+    override fun lightTrough(data: Int) = (-1).toByte()
 
     override fun connectStage(terrain: TerrainClient,
                               x: Int,
